CT-163 useHandleResize Hook 구현 및 TopBar PopUp 위치 수정
